ai_suite/requirements.txt
# Web & API
fastapi
uvicorn
httpx
python-multipart

# Data Processing
pandas
openpyxl
numpy
matplotlib

# Document Generation
python-docx
fpdf

# Machine Learning Frameworks
scikit-learn
joblib        # For loading .pkl model files
# xgboost     # Uncomment if using xgboost (Will increase build time)
# tensorflow  # Uncomment if you plan to use Deep Learning (Warning: Large)
# torch       # Uncomment if you plan to use PyTorch (Warning: Large)

# Utilities
importlib-metadata
python-magic